Synchronization feature
1) Personalities for pub/sub.
It allows to use a separate exchange/queue_names when working with a
different personality. In our case, we use it with the "sync"
personality.

2) Split the original rake task
Instead of having one process for both publisher worker and subscriber
worker, we have one for each. This is important when it comes to
working with different personalities.

3) It has a progress bar